v0.8.2 release
Fix problem that could cause the macro to crash when used for the first time in a drawing.
Added a limit of 32,750 entities for shadow processing (processing was previously unreliable when the shadow count reached 32,768)
changes to way the macro refreshes the screen (to make it faster and better).
Stay in selection mode until S0 is pressed when adding or removing entities from a selection set in the Shadow Entities menu.
Minor updates to some label messages to make them more informative.